calcs-api-code/pyproject.toml
[project]
name = "calcs-api-code"
version = "0.1.0"
description = "Direct Python client for Calcs API - designed for code execution patterns"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["calcs_api_code"]
